2020 8.29

+++ article/views.py
@@ -4,6 +4,7 @@ from .models import ArticlePost
 import markdown
 from django.contrib.auth.models import User
 from .forms import ArticlePostForm
+from django.contrib.auth.decorators import login_required
 
 
 # Create your views here.
@@ -33,6 +34,7 @@ def article_detail(request, id):
     return render(request, 'article/detail.html', context)
 
 
+@login_required(login_url='/userprofile/login/')
 def article_create(request):
     # 写文章的视图
     if request.method == 'POST':
@@ -40,10 +42,7 @@ def article_create(request):
         if article_post_form.is_valid():
             # 保存数据，但暂时不提交到数据库中
             new_article = article_post_form.save(commit=False)
-            # 指定数据库中 id=1 的用户为作者
-            # 如果你进行过删除数据表的操作，可能会找不到id=1的用户
-            # 此时请重新创建用户，并传入此用户的id
-            new_article.author = User.objects.get(id=1)
+            new_article.author = User.objects.get(id=request.user.id)
             # 将新文章保存到数据库中
             new_article.save()
             # 完成后返回到文章列表
@@ -60,7 +59,12 @@ def article_create(request):
         return render(request, 'article/create.html', context)
 
 
+@login_required(login_url='/userprofile/login/')
 def article_safe_delete(request, id):
+    author_id = ArticlePost.objects.get(id=id).author_id
+    user_id = request.user.id
+    if author_id != user_id:
+        return HttpResponse("你没有权限修改此用户信息。")
     if request.method == 'POST':
         article = ArticlePost.objects.get(id=id)
         article.delete()
@@ -69,6 +73,7 @@ def article_safe_delete(request, id):
         return HttpResponse("仅允许post请求")
 
 
+@login_required(login_url='/userprofile/login/')
 def article_update(request, id):
     """
     更新文章的视图函数
@@ -76,9 +81,12 @@ def article_update(request, id):
     GET方法进入初始表单页面
     id： 文章的 id
     """
-
     # 获取需要修改的具体文章对象
     article = ArticlePost.objects.get(id=id)
+    author_id = article.author_id
+    user_id = request.user.id
+    if author_id != user_id:
+        return HttpResponse("你没有权限修改此用户信息。")
     # 判断用户是否为 POST 提交表单数据
     if request.method == "POST":
         # 将提交的数据赋值到表单实例中
@@ -94,7 +102,6 @@ def article_update(request, id):
         # 如果数据不合法，返回错误信息
         else:
             return HttpResponse("表单内容有误，请重新填写。")
-
     # 如果用户 GET 请求获取数据
     else:
         # 创建表单类实例
